+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixes
 
-- Fix sorting prerelease versions ([#78](https://github.com/getsentry/github-workflows/pull/78))
+- Don't update from a manually-updated prerelease to a latest stable release that is earlier than the prerelease ([#78](https://github.com/getsentry/github-workflows/pull/78))
 
 ## 2.11.0