+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Don't update from a manually-updated prerelease to a latest stable release that is earlier than the prerelease ([#78](https://github.com/getsentry/github-workflows/pull/78))
 
+### Dependencies
+
+- Bump WORKFLOW-TEST-DEPENDENCY-DO-NOT-MERGE from v2.0.0 to v2.0.4 ([#66](https://github.com/getsentry/github-workflows/pull/66))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#204)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.0.0...2.0.4)
+
 ## 2.11.0
 
 ### Features