+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Don't update from a manually-updated prerelease to a latest stable release that is earlier than the prerelease ([#78](https://github.com/getsentry/github-workflows/pull/78))
+- Cross-repo links in changelog notes ([#82](https://github.com/getsentry/github-workflows/pull/82))
 
 ## 2.11.0